% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Draw_a_name}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{the name you want to draw}

\item{the_sex}{}
}
\value{
a graph
}
\description{
Draw_a_name
}
\examples{
draw_a_name("Guilhem","M")
}
