% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Draw_names}
\usage{
draw_names(x)
}
\arguments{
\item{x}{list of the names you want to draw}

\item{the_sex}{}
}
\value{
a graph
}
\description{
Draw_names
}
\examples{
draw_names(c("Louis","Paul"))
}
