% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries.R
\name{fisheries}
\alias{fisheries}
\title{Describe the most frequently caught fish for each location and total revenue for fisheries}
\usage{
fisheries(fish_price, fish_catches, showplot = FALSE)
}
\arguments{
\item{fish_price}{a table that has prices for different fish}

\item{fish_catches}{a table that has the number caught for each fish species for each location}

\item{showplot}{if requested, a graph of revenue by location and total revenue}
}
\value{
list with the following items
\describe{
\item{mostfrequent}{ Most frequently caught fish in each location}
\item{totalrevenue_bylocation}{Total revenue for each location}
\item{totalrevenue}{Total fisheries revenue sum}
}
}
\description{
Describe the most frequently caught fish for each location and total revenue for fisheries
}
\examples{
fisheries(fish_price, fish_catches, showplot=TRUE)
}
\author{
Celine Mol and Paige FitzGibbon
}
