% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishgrowth.R
\name{fishgrowth}
\alias{fishgrowth}
\title{Predicts fish growth rate based on temperature}
\usage{
fishgrowth(a, b, c, d, T)
}
\arguments{
\item{a}{zero-order coefficient}

\item{b}{first-order coefficient}

\item{c}{second-order coefficient}

\item{d}{third-order coefficient}

\item{T}{temperature of fishery in degrees celsius}
}
\description{
Predicts fish growth rate based on temperature
}
\examples{
fishgrowth(0.1,0.2,0.3,0.4, T=seq(from = 0, to = 18, by = 2))
}
\references{
https://www.sciencedirect.com/science/article/abs/pii/S0044848607005182
}
\author{
Celine Mol and Paige FitzGibbon
}
