% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishmarketdata.R
\docType{data}
\name{fishmarketdata}
\alias{fishmarketdata}
\title{Fisheries Example Data}
\format{A data frame with 6 rows and 3 variables. The rows represent different type of fish caught: "Cod", "Salmon", "Steelhead", "Tuna", "Shark", "Sole".
\itemize{
\item fish
\item location1
\item location2
\item location3
}}
\source{
Celine Mol
}
\usage{
fishmarketdata
}
\description{
Data that was randomly generated as example data for number of each type of fish caught in each location.
}
\keyword{datasets}
