% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear}
\alias{linear}
\title{linear: Fitting Simple Linear Regression Models}
\usage{
linear(formula.lm, data)
}
\arguments{
\item{formula.lm}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.}

\item{data}{an optional data frame, list or environment (or object coercible by \code{\link{as.data.frame}} to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which linear is called.}
}
\value{

}
\description{
`linear` is used to fit simple linear regression models. And it can also returns 95% confidence intervals for one or more parameters in a fitted model.
}
\details{
Models for linear are specified symbolically. A typical model has the form `response ~ terms` where `response` is the (numeric) response vector and `terms` is a series of terms which specifies a linear predictor for `response`.

A formula has an implied intercept term. To remove this use either `y ~ x - 1` or `y ~ 0 + x`.
}
\examples{
attach(mydata) #using the data
cc<-Depression ~ conFatalism + Sex + R_E + factor(Age1)
results<-linear(cc,mydata)
coef_c<-results$Coefficients # the coefficients and its related values.
coef_c
results$F.value
## input"formula.lm" must be a formula.
\dontrun{
attach(mydata)
   linear(Sex,mydata)
}

}
\references{
LM: Fitting linear models (1969) RDocumentation. R-core R-core@R-project.org. Available at: https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/lm (Accessed: November 20, 2022).
}
