% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatChange.R
\name{MatChange}
\alias{MatChange}
\title{Matrix Changer}
\usage{
MatChange(A, x)
}
\arguments{
\item{A}{A square matrix of length n X n}

\item{x}{a vector of length n}
}
\value{
an altered matrix
}
\description{
This function takes in a matrix and a vector and multiplies
the transpose of the vector by the inverse of the matrix by the
original, non-transposed vector.
}
