% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatStandard.R
\name{MatStandard}
\alias{MatStandard}
\title{Standardization of Matrix Columns}
\usage{
MatStandard(A)
}
\arguments{
\item{A}{the matrix to be standardized}
}
\value{
new standardized matrix
}
\description{
This Function standardizes all column in a matrix by subtracting
the mean of all values in a column divided by its standardization
from each value in the column
}
