% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_excel.R
\name{multi_excel}
\alias{multi_excel}
\title{multi_excel_reader_function}
\usage{
multi_excel(file_path)
}
\arguments{
\item{file_path}{The path of your file}
}
\value{
a dataframe
}
\description{
multi_excel_reader_function
}
