% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSd.R
\name{MVSd}
\alias{MVSd}
\title{Mean, Variation, and Standard Deviation on a numeric vector}
\usage{
MVSd(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
mean, variation, and standard deviation in a list in that order
}
\description{
The User will enter in a numeric Vector and the function
will return the mean, variation, and standard deviation of
that vector in that order.
}
