% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSd_Inequal.R
\name{MVSd_Inequal}
\alias{MVSd_Inequal}
\title{Mean, Variation, and Standard Deviation on a numeric vector
given unequal probabilities of data values}
\usage{
MVSd_Inequal(x, p)
}
\arguments{
\item{x}{A numeric vector}

\item{p}{probabilities of the corresponding x}
}
\value{
mean, variation, and standard deviation in a list in that order
}
\description{
The User will enter two numeric Vectors of the same size, on being tha values
and the other being the probability of those corresponding value. 
The function will return the mean, variation, and standard deviation of
that vector in that order.
}
