% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{my_confint}
\alias{my_confint}
\title{Confidence Interval}
\usage{
my_confint(formula, data, level = 0.95)
}
\arguments{
\item{formula}{input formula such as y ~ x1 + x2}

\item{data}{input data such as mtcars}

\item{level}{numerical value, the significance of the CI, default = 0.95}
}
\value{
The upper and lower bounds of confidence interval: CI_matrix
}
\description{
Confidence Interval
}
\examples{
data = mtcars
formula = mpg ~ disp + hp
my_confint(formula, data, level = 0.95)
}
