% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{my_csv_reader}
\usage{
my_csv_reader(file_path, sep = x, dev = y)
}
\arguments{
\item{file_path}{the path of your file}

\item{sep}{the separator}

\item{dev}{the decimal}
}
\value{
reads the csv
}
\description{
my_csv_reader
}
