% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{my_lm}
\alias{my_lm}
\title{Multiply Linear Regression}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{input formula such as y ~ x1 + x2}

\item{data}{input data such as mtcars}
}
\value{
List containing the following elements
\itemize{
\item beta estimates(coefficients)
\item fitted_values
\item residuals
\item mse
\item rsquared
\item adj_rsquared
\item SE
\item t
\item p.value
\item F_stat
\item df1
\item df2
\item CI
}
}
\description{
Multiply Linear Regression
}
\examples{
data = mtcars
formula = mpg ~ disp + hp
my_lm(formula, data)
}
