% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myApply.R
\name{myApply}
\alias{myApply}
\title{A simple homemade apply function}
\usage{
myApply(X, MARGIN, FUN, ...)
}
\arguments{
\item{X}{a matrix to be applied on}

\item{MARGIN}{for which dimension will be applied on}

\item{FUN}{the function that will change the matrix by}

\item{...}{additional paramerters for the function if neccessary}
}
\value{
the results of the permutated matrix in simple array form
}
\description{
Changes a user supplied matrix by a user supplied function on the dimension
that the user chooses with 1 being the first dimension and 2 being the seconnd
It also allows user to insert ther own further parameters
}
