% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{save_as_csv}
\usage{
save_as_csv(x, name)
}
\arguments{
\item{x}{a dataframe}

\item{name}{name of the csv you want to write}
}
\value{
creates a csv file from a dataframe
}
\description{
save_as_csv
}
\examples{
save_as_csv(iris,"iris3.csv")
}
