% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_AIC.R
\name{calculate_aic}
\alias{calculate_aic}
\title{Calculate the AIC manually}
\usage{
calculate_aic(formula, data)
}
\arguments{
\item{data}{}
}
\value{
List containing the following elements
\itemize{
\item beta estimates(coefficients)
\item fitted_values
\item residuals
\item rsquared
}
}
\description{
Calculate the AIC manually
}
\examples{
data = mtcars
formula = mpg ~ disp + hp
calculate_aic(formula, data)
}
