% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries.R
\name{fisheries}
\alias{fisheries}
\title{Describe diversity based on a list of species}
\usage{
fisheries(fish_price, fish_catches, showplot = FALSE)
}
\arguments{
\item{fish_price}{a table that has prices for different fish}

\item{fish_catches}{a table that has the number caught for each fish species for each location}
}
\value{
list with the following items
\describe{
\item{mostfrequent}{ Most frequently caught fish in each location}
\item{totalrevenue_bylocation}{Total revenue for each location}
\item{totalrevenue}{Total fisheries revenue sum}
}
}
\description{
Compute a species diversity index
}
\examples{
computediversity(c("butterfly","butterfly","mosquito","butterfly","ladybug","ladybug")))
}
\references{
http://www.tiem.utk.edu/~gross/bioed/bealsmodules/simpsonDI.html
}
