% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishgrowth.R
\name{fishgrowth}
\alias{fishgrowth}
\title{Predicts fish growth rate based on temperature}
\usage{
fishgrowth(a, b, c, d, T)
}
\arguments{
\item{a}{{zero order parameter}}

\item{b}{{first order parameter}}

\item{c}{{second order parameter}}

\item{d}{{third order parameter}}

\item{T}{temperature of fishery in degrees celsius}
}
\value{
G = a + bT + cT^2 + dT^3
}
\description{
Predicts fish growth rate based on temperature
}
\examples{
fishgrowth(a = -0.4970, b = 0.1656, c = 0.08588, d = -0.004266, T = 14.3)
fishgrowth(0.1,0.2,0.3,0.4, T=seq(from = 0, to = 18, by = 2))
}
\references{
https://www.sciencedirect.com/science/article/abs/pii/S0044848607005182
}
\author{
Celine Mol and Paige FitzGibbon
}
