% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishgrowth.R
\name{fishgrowth}
\alias{fishgrowth}
\title{Predicts fish growth rate based on temperature}
\usage{
fishgrowth(a, b, c, d, T)
}
\arguments{
\item{T}{temperature of fishery in degrees celsius}

\item{coefficients}{\describe{
\item {a} {zero order parameter}
\item {b} {first order parameter}
\item {c} {second order parameter}
\item{d}{third order parameter}
}}
}
\description{
Predicts fish growth rate based on temperature
}
\references{
https://www.sciencedirect.com/science/article/abs/pii/S0044848607005182
}
