% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_AIC.R
\name{my_lm}
\alias{my_lm}
\title{Multiply Linear Regression}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{11}

\item{data}{1111}
}
\value{
List containing the following elements
\itemize{
\item beta estimates(coefficients)
\item fitted_values
\item residuals
\item rsquared
}
}
\description{
Multiply Linear Regression
}
\examples{
data = mtcars
formula = mpg ~ disp + hp
my_lm(formula, data)
}
